do some crazy contortions with getting paths correct, and add some automated regression tests in the process :0

+++ tests/rmdir_test/rmdir_test.py
@@ -1,5 +1,12 @@
 import os
 
+if os.path.isdir('/tmp/guinea-pig-dir/guinea-pig-subdir'):
+  os.rmdir('/tmp/guinea-pig-dir/guinea-pig-subdir')
+
+if os.path.isdir('/tmp/guinea-pig-dir'):
+  os.rmdir('/tmp/guinea-pig-dir')
+
 os.mkdir('/tmp/guinea-pig-dir')
-os.rmdir('/tmp/guinea-pig-dir')
+os.mkdir('/tmp/guinea-pig-dir/guinea-pig-subdir')
+os.rmdir('/tmp/guinea-pig-dir/guinea-pig-subdir')
